refactor(affiliate): modularize referral modal content
